requirements.txt
accelerate==0.32.1
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
argcomplete==3.5.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.15.0
beautifulsoup4==4.12.3
bitsandbytes==0.43.3
bleach==6.1.0
blessed==1.20.0
boto==2.49.0
cachetools==5.4.0
cattrs==23.1.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.0
comm==0.2.2
compressed-tensors==0.6.0
contourpy==1.2.1
crcmod==1.7
cryptography==43.0.0
cycler==0.12.1
datasets==3.1.0
debugpy==1.8.2
decorator==5.1.1
decord==0.6.0
deepspeed==0.14.4
defusedxml==0.7.1
diffusers==0.29.2
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docstring-to-markdown==0.12
einops==0.8.0
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.115.4
fasteners==0.19
fastjsonschema==2.20.0
filelock==3.15.4
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
gcs-oauth2-boto-plugin==3.2
gdown==5.2.0
gguf==0.10.0
google-apitools==0.5.32
google-auth==2.17.0
google-auth-httplib2==0.2.0
google-reauth==0.1.1
gpustat==1.1.1
gsutil==5.30
h11==0.14.0
hjson==3.1.0
httpcore==1.0.5
httplib2==0.20.4
httptools==0.6.4
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
importlib-metadata==6.8.0
inquirerpy==0.3.4
interegular==0.3.3
ipykernel==6.29.5
ipython==8.26.0
isoduration==20.11.0
jedi==0.19.1
jedi-language-server==0.41.1
Jinja2==3.1.4
jiter==0.7.0
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
kiwisolver==1.4.5
lark==1.2.2
llvmlite==0.43.0
lm-format-enforcer==0.10.6
lsprotocol==2023.0.0b1
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
microvenv==2023.5.post1
mistral_common==1.4.4
mistune==3.0.2
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.0.5
multiprocess==0.70.16
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.161
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.2
oauth2client==4.1.3
openai==1.54.3
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
outlines==0.0.46
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.3
partial-json-parser==0.2.1.1.post4
pexpect==4.9.0
pfzy==0.3.4
pillow==10.4.0
platformdirs==4.2.2
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==5.28.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==18.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycountry==24.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pygls==1.0.2
Pygments==2.18.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
pyu2f==0.1.5
PyYAML==6.0.1
pyzmq==26.0.3
qwen-vl-utils==0.0.4
ray==2.38.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
retry_decorator==1.1.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.19.0
rsa==4.7.2
safetensors==0.4.3
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.41.2
sympy==1.13.0
tenacity==9.0.0
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==1.0.9
tinycss2==1.3.0
tokenizers==0.20.3
tomli==2.0.1
torch==2.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.46.2
transformers-stream-generator==0.0.5
triton==3.0.0
typeguard==3.0.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.32.0
uvloop==0.21.0
vllm==0.6.3.post1
watchfiles==0.24.0
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.0
xformers==0.0.27.post2
xxhash==3.5.0
yarl==1.9.4
zipp==3.17.0
